refactor: fixed text size consistency
