+++ fableous-fe/src/components/canvas/CanvasToolbar.tsx
@@ -16,6 +16,7 @@ import { proto as pb } from "../../proto/message_pb";
 import BrushWidthIcon from "./BrushWidthIcon";
 import CanvasToolbarTooltip from "./CanvasToolbarTooltip";
 import { TutorialTargetId } from "../../tutorialTargetIds";
+import { BRUSH_COLORS, BRUSH_WIDTHS } from "./constants";
 
 interface CanvasToolbarProps {
   role: pb.ControllerRole;
@@ -28,22 +29,7 @@ interface CanvasToolbarProps {
   setToolNormWidth: React.Dispatch<React.SetStateAction<number>>;
 }
 
-export const COLORS = [
-  "#282828FF",
-  "#A4C421FF",
-  "#34AF2EFF",
-  "#34A59CFF",
-  "#234EB0FF",
-  "#3C2DADFF",
-  "#613494FF",
-  "#A3238FFF",
-  "#E41728FF",
-  "#EE652EFF",
-  "#F49129FF",
-  "#F9B72BFF",
-];
 const ERASE_COLOR = "#00000000";
-export const BRUSH_WIDTHS = [1 / 64, 2 / 64, 3 / 64, 4 / 64, 5 / 64];
 const ICON_STROKE_WIDTH_RATIO = 128;
 
 const useStyles = makeStyles({
@@ -212,7 +198,7 @@ const CanvasToolbar = forwardRef<ImperativeCanvasRef, CanvasToolbarProps>(
                         display: "flex",
                       }}
                     >
-                      {COLORS.map((color) => (
+                      {BRUSH_COLORS.map((color) => (
                         <Button
                           component="div"
                           onClick={() => {
