+++ fableous-fe/src/components/canvas/constants.ts
@@ -1,3 +1,22 @@
 export const ASPECT_RATIO = 9 / 16;
 export const SCALE = 2;
 export const SELECT_PADDING = 12;
+
+export const BRUSH_COLORS = [
+  "#282828FF",
+  "#A4C421FF",
+  "#34AF2EFF",
+  "#34A59CFF",
+  "#234EB0FF",
+  "#3C2DADFF",
+  "#613494FF",
+  "#A3238FFF",
+  "#E41728FF",
+  "#EE652EFF",
+  "#F49129FF",
+  "#F9B72BFF",
+];
+export const BRUSH_WIDTHS = [1 / 64, 2 / 64, 3 / 64, 4 / 64, 5 / 64];
+
+export const TEXT_FONTSIZE = 0.6 / 64;
+export const TEXT_COLOR = "#101010FF";
